# feat: improve header validation speed